+++ seed/csharp-sdk/basic-auth-environment-variables/src/SeedBasicAuthEnvironmentVariables/SeedBasicAuthEnvironmentVariablesClient.cs
@@ -1,4 +1,3 @@
-using System;
 using SeedBasicAuthEnvironmentVariables.Core;
 
 #nullable enable
@@ -23,11 +22,22 @@ public partial class SeedBasicAuthEnvironmentVariablesClient
             "PASSWORD",
             "Please pass in password or set the environment variable PASSWORD."
         );
-        _client = new RawClient(
-            new Dictionary<string, string>() { { "X-Fern-Language", "C#" } },
-            new Dictionary<string, Func<string>>(),
-            clientOptions ?? new ClientOptions()
+        var defaultHeaders = new Headers(
+            new Dictionary<string, string>()
+            {
+                { "X-Fern-Language", "C#" },
+                { "User-Agent", "Fernbasic-auth-environment-variables/0.0.1" },
+            }
         );
+        clientOptions ??= new ClientOptions();
+        foreach (var header in defaultHeaders)
+        {
+            if (!clientOptions.Headers.ContainsKey(header.Key))
+            {
+                clientOptions.Headers[header.Key] = header.Value;
+            }
+        }
+        _client = new RawClient(clientOptions);
         BasicAuth = new BasicAuthClient(_client);
         Errors = new ErrorsClient(_client);
     }