+++ seed/csharp-sdk/idempotency-headers/src/SeedIdempotencyHeaders/SeedIdempotencyHeadersClient.cs
@@ -1,4 +1,3 @@
-using System;
 using SeedIdempotencyHeaders.Core;
 
 #nullable enable
@@ -11,15 +10,23 @@ public partial class SeedIdempotencyHeadersClient
 
     public SeedIdempotencyHeadersClient(string? token = null, ClientOptions? clientOptions = null)
     {
-        _client = new RawClient(
+        var defaultHeaders = new Headers(
             new Dictionary<string, string>()
             {
                 { "Authorization", $"Bearer {token}" },
                 { "X-Fern-Language", "C#" },
-            },
-            new Dictionary<string, Func<string>>(),
-            clientOptions ?? new ClientOptions()
+                { "User-Agent", "Fernidempotency-headers/0.0.1" },
+            }
         );
+        clientOptions ??= new ClientOptions();
+        foreach (var header in defaultHeaders)
+        {
+            if (!clientOptions.Headers.ContainsKey(header.Key))
+            {
+                clientOptions.Headers[header.Key] = header.Value;
+            }
+        }
+        _client = new RawClient(clientOptions);
         Payment = new PaymentClient(_client);
     }