+++ seed/csharp-sdk/pagination/src/SeedPagination/SeedPaginationClient.cs
@@ -1,4 +1,3 @@
-using System;
 using SeedPagination.Core;
 
 #nullable enable
@@ -11,11 +10,22 @@ public partial class SeedPaginationClient
 
     public SeedPaginationClient(string token, ClientOptions? clientOptions = null)
     {
-        _client = new RawClient(
-            new Dictionary<string, string>() { { "X-Fern-Language", "C#" } },
-            new Dictionary<string, Func<string>>(),
-            clientOptions ?? new ClientOptions()
+        var defaultHeaders = new Headers(
+            new Dictionary<string, string>()
+            {
+                { "X-Fern-Language", "C#" },
+                { "User-Agent", "Fernpagination/0.0.1" },
+            }
         );
+        clientOptions ??= new ClientOptions();
+        foreach (var header in defaultHeaders)
+        {
+            if (!clientOptions.Headers.ContainsKey(header.Key))
+            {
+                clientOptions.Headers[header.Key] = header.Value;
+            }
+        }
+        _client = new RawClient(clientOptions);
         Users = new UsersClient(_client);
     }