+++ seed/csharp-sdk/response-property/src/SeedResponseProperty/SeedResponsePropertyClient.cs
@@ -1,4 +1,3 @@
-using System;
 using SeedResponseProperty.Core;
 
 #nullable enable
@@ -11,11 +10,22 @@ public partial class SeedResponsePropertyClient
 
     public SeedResponsePropertyClient(ClientOptions? clientOptions = null)
     {
-        _client = new RawClient(
-            new Dictionary<string, string>() { { "X-Fern-Language", "C#" } },
-            new Dictionary<string, Func<string>>(),
-            clientOptions ?? new ClientOptions()
+        var defaultHeaders = new Headers(
+            new Dictionary<string, string>()
+            {
+                { "X-Fern-Language", "C#" },
+                { "User-Agent", "Fernresponse-property/0.0.1" },
+            }
         );
+        clientOptions ??= new ClientOptions();
+        foreach (var header in defaultHeaders)
+        {
+            if (!clientOptions.Headers.ContainsKey(header.Key))
+            {
+                clientOptions.Headers[header.Key] = header.Value;
+            }
+        }
+        _client = new RawClient(clientOptions);
         Service = new ServiceClient(_client);
     }